requirements.txt
Flask
Redis
Flask-SQLAlchemy
Flask-WTF
py4j
Spark
Celery
Flask-Webpack
gunicorn
python-dotenv
facepy
jedi
service_factory
